// frontend/src/api/generated/models/sportType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type SportType = typeof SportType[keyof typeof SportType];


export const SportType = {
  Running: 'Running',
  Cycling: 'Cycling',
  Swimming: 'Swimming',
  Strength: 'Strength',
  CrossTraining: 'CrossTraining',
  Rest: 'Rest',
  Other: 'Other',
} as const;