// frontend/src/api/generated/models/workoutTemplateDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */
import type { SportType } from './sportType';
import type { WorkoutSegmentDto } from './workoutSegmentDto';

export interface WorkoutTemplateDto {
  id?: string;
  coachUserId?: string;
  /** @nullable */
  name?: string | null;
  sport?: SportType;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  segments?: WorkoutSegmentDto[] | null;
}